Finished mvp gallery & switching to testing
