+++ client/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Gallery from './gallery';
+import styled from 'styled-components';
+
+ReactDOM.render(<Gallery />, document.getElementById('gallery'))
