+++ package.json
@@ -4,9 +4,13 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest",
+    "test": "jest --watch",
     "build": "webpack --watch",
     "start": "nodemon ./server/server.js"
+  },
+  "jest": {
+    "snapshotSerializers": ["enzyme-to-json/serializer"],
+    "setupFiles": ["./setupTests.js"]
   },
   "repository": {
     "type": "git",
@@ -26,18 +30,27 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "express": "^4.16.2",
-    "jest": "^22.3.0",
     "mongodb": "^3.1.1",
     "mongoose": "^5.2.4",
     "morgan": "^1.9.0",
     "mysql": "^2.16.0",
     "nodemon": "^1.18.3",
     "react": "^16.2.0",
+    "react-addons-test-utils": "^15.6.2",
     "react-dom": "^16.2.0",
+    "react-test-renderer": "^16.4.1",
     "styled-components": "^3.3.3",
     "webpack": "^3.12.0"
   },
   "devDependencies": {
+    "babel-jest": "^23.4.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
+    "babel-polyfill": "^6.26.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.4",
+    "jest": "^22.4.4",
+    "jest-enzyme": "^6.0.2",
     "node-sass": "^4.9.2",
     "sass-loader": "^7.0.3"
   }
